+++ src/searchRequest.js
@@ -384,6 +384,7 @@ var filterPassThroughMethods = [
 	'whereExists', 'orWhereExists', 'whereNotExists', 'orWhereNotExists',
 	'whereIn', 'orWhereIn', 'whereNotIn', 'orWhereNotIn',
 	'whereLike', 'orWhereLike', 'whereNotLike', 'orWhereNotLike',
+	'whereRegex', 'orWhereRegex', 'whereNotRegex', 'orWhereNotRegex',
 	'getFilter', 'getFilterValue'
 ];