+++ changelog.md
@@ -1,5 +1,6 @@
 ## Changelog
 
+- Added regex filter operator
 - Updated readme to include word operator companion methods.
 - Added helper methods for `like` and `not like` filters.